+++ packages/composites/proof-of-reserves/src/utils/balance.ts
@@ -43,13 +43,13 @@ export const adaptersV2: v2AdapterImplementation[] = [
   avalanchePlatform as unknown as v2AdapterImplementation,
 ]
 
-export const adaptersV3: v3AdapterImplementation[] = [polkadotBalance as v3AdapterImplementation]
-
-export type Indexer = typeof adaptersV2[number]['NAME'] & typeof adaptersV3[number]['name']
+export const adaptersV3: v3AdapterImplementation[] = [
+  polkadotBalance as unknown as v3AdapterImplementation,
+]
 
 // Get balances for address set
 export const runBalanceAdapter = async (
-  indexer: Indexer,
+  indexer: string,
   context: AdapterContext,
   confirmations: number,
   config: Config,