+++ packages/composites/proof-of-reserves/src/utils/protocol.ts
@@ -36,10 +36,8 @@ export const adaptersV2: v2AdapterImplementation[] = [
 ]
 
 export const adaptersV3: v3AdapterImplementation[] = [
-  moonbeamAddressList as v3AdapterImplementation,
+  moonbeamAddressList as unknown as v3AdapterImplementation,
 ]
-
-export type Protocol = typeof adaptersV2[number]['NAME'] & typeof adaptersV3[number]['name']
 
 type AddressData = { token: string; chainId: string; network: string } | AddressList
 
@@ -53,7 +51,7 @@ type AddressObject = { address: string; network: string; chainId: string }
 export const runProtocolAdapter = async (
   jobRunID: string,
   context: AdapterContext,
-  protocol: Protocol,
+  protocol: string,
   data: AddressData,
   config: Config,
 ): Promise<AdapterResponse> => {