+++ packages/sources/cfbenchmarks-test/src/index.ts
@@ -1,15 +1,14 @@
 import { expose, ServerInstance } from '@chainlink/external-adapter-framework'
-import { PriceAdapter } from '@chainlink/external-adapter-framework/adapter'
+import { Adapter } from '@chainlink/external-adapter-framework/adapter'
 import { customSettings } from './config'
-import { crypto } from './endpoint'
-import { requestTransforms } from './endpoint/common/crypto'
+import { crypto, birc } from './endpoint'
+import { CryptoPriceEndpoint } from '@chainlink/external-adapter-framework/adapter'
 
-export const adapter = new PriceAdapter({
+export const adapter = new Adapter({
   name: 'CFBENCHMARKS',
-  endpoints: [crypto],
+  endpoints: [crypto as CryptoPriceEndpoint<any>, birc],
   defaultEndpoint: crypto.name,
   customSettings,
-  requestTransforms,
 })
 
 export const server = (): Promise<ServerInstance | undefined> => expose(adapter)