+++ packages/sources/cfbenchmarks/src/endpoint/index.ts
@@ -1,5 +1,7 @@
 import type { TInputParameters as ValuesInputParameters } from './values'
+import type { TInputParameters as BircInputParameters } from './birc'
 
-export type TInputParameters = ValuesInputParameters
+export type TInputParameters = ValuesInputParameters | BircInputParameters
 
 export * as values from './values'
+export * as birc from './birc'