+++ packages/sources/intrinio-test/src/config/index.ts
@@ -10,9 +10,4 @@ export const customSettings = {
     require: true,
     sensitive: true,
   },
-  WS_ENABLED: {
-    description: 'Whether data should be returned from websocket or not',
-    type: 'boolean',
-    default: false,
-  },
 } as const