+++ packages/sources/oanda/src/transport/modifiedSSE.ts
@@ -62,7 +62,7 @@ export class ModifiedSseTransport<T extends TransportGenerics> extends Streaming
     config: AdapterConfig<T['CustomSettings']>,
     endpointName: string,
   ): Promise<void> {
-    super.initialize(dependencies, config, endpointName)
+    super.initialize(dependencies, config, endpointName, 'sse')
   }
 
   async streamHandler(
@@ -88,7 +88,7 @@ export class ModifiedSseTransport<T extends TransportGenerics> extends Streaming
 
           const stream = response?.data
 
-          const eventHandlerGenerator = (listener: typeof this.config.eventListeners[0]) => {
+          const eventHandlerGenerator = (listener: (typeof this.config.eventListeners)[0]) => {
             return (event: MessageEvent) => {
               const providerDataReceivedUnixMs = Date.now()
 
@@ -103,7 +103,7 @@ export class ModifiedSseTransport<T extends TransportGenerics> extends Streaming
                 }
                 return result
               })
-              this.responseCache.write(results)
+              this.responseCache.write('sse', results)
             }
           }