+++ packages/sources/tradermade-test/src/index.ts
@@ -2,7 +2,6 @@ import { expose, ServerInstance } from '@chainlink/external-adapter-framework'
 import { PriceAdapter } from '@chainlink/external-adapter-framework/adapter'
 import { customSettings } from './config'
 import includes from './config/includes.json'
-import overrides from './config/overrides.json'
 import { forex, live } from './endpoint'
 
 export const adapter = new PriceAdapter({
@@ -11,7 +10,6 @@ export const adapter = new PriceAdapter({
   customSettings,
   endpoints: [forex, live],
   includes,
-  overrides: overrides.tradermade,
   rateLimiting: {
     tiers: {
       basic: {