+++ packages/composites/bsol-price/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @chainlink/bsol-price-adapter
 
+## 2.2.37
+
+### Patch Changes
+
+- 65014014d: Upgraded typescript version to 4.9.5
+- Updated dependencies [b29509be0]
+- Updated dependencies [65014014d]
+  - @chainlink/ea-bootstrap@2.26.0
+  - @chainlink/ea-test-helpers@1.4.2
+  - @chainlink/token-allocation-adapter@1.9.3
+  - @chainlink/solana-view-function-adapter@2.2.28
+
 ## 2.2.36
 
 ### Patch Changes