+++ packages/sources/alongside/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @chainlink/alongside-adapter
 
+## 1.1.1
+
+### Patch Changes
+
+- 65014014d: Bumped v3 framework version. Upgraded typescript to version 4.9.5
+
+## 1.1.0
+
+### Minor Changes
+
+- 8e6e562f8: Updated to allow generic rpc provider
+
 ## 1.0.0
 
 ### Major Changes