+++ packages/sources/coinmetrics-test/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @chainlink/coinmetrics-test-adapter
 
+## 1.0.8
+
+### Patch Changes
+
+- 65014014d: Upgraded typescript version to 4.9.5
+
 ## 1.0.7
 
 ### Patch Changes