+++ packages/sources/dar/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @chainlink/dar-adapter
 
+## 1.0.6
+
+### Patch Changes
+
+- 65014014d: Bumped v3 framework version. Upgraded typescript to version 4.9.5
+
 ## 1.0.5
 
 ### Patch Changes