+++ packages/sources/lotus/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @chainlink/lotus-adapter
 
+## 2.2.28
+
+### Patch Changes
+
+- 65014014d: Upgraded typescript version to 4.9.5
+- Updated dependencies [b29509be0]
+- Updated dependencies [65014014d]
+  - @chainlink/ea-bootstrap@2.26.0
+  - @chainlink/json-rpc-adapter@1.3.28
+
 ## 2.2.27
 
 ### Patch Changes